refactor: move auth components and hooks from packages/features to apps/web/modules
