+++ apps/web/app/(use-page-wrapper)/apps/routing-forms/[...pages]/RoutingFormAuthGuard.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { useRedirectToLoginIfUnauthenticated } from "@calcom/features/auth/lib/hooks/useRedirectToLoginIfUnauthenticated";
+import { useRedirectToLoginIfUnauthenticated } from "@calcom/web/modules/auth/hooks/useRedirectToLoginIfUnauthenticated";
 
 export function RoutingFormAuthGuard({ children }: { children: React.ReactNode }) {
   useRedirectToLoginIfUnauthenticated();
