+++ apps/web/modules/auth/hooks/useRedirectToLoginIfUnauthenticated.tsx
