+++ apps/web/modules/bookings/components/Booker.test.tsx
@@ -32,7 +32,7 @@ vi.mock("next/navigation", async (importOriginal) => {
 });
 
 import "@calcom/dayjs/__mocks__";
-import "@calcom/features/auth/Turnstile";
+import "@calcom/web/modules/auth/components/Turnstile";
 
 import { render, screen } from "@calcom/features/bookings/Booker/__tests__/test-utils";
 import type { BookerProps, WrappedBookerProps } from "@calcom/features/bookings/Booker/types";
